Rename Package and Project
Đổi tên package cho hợp lí
Đổi tên package trong tệp struts.xml, pom.xml và web.xml
