+++ .settings/org.eclipse.wst.common.component
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
-    <wb-module deploy-name="struts2_crud_mybatis-0.0.1-SNAPSHOT">
+    <wb-module deploy-name="dashboard-0.0.1-SNAPSHOT">
         <wb-resource deploy-path="/" source-path="/target/m2e-wtp/web-resources"/>
         <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/java"/>
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
-        <property name="context-root" value="struts2_crud_mybatis"/>
-        <property name="java-output-path" value="/struts2_crud_mybatis/target/classes"/>
+        <property name="context-root" value="dashboard"/>
+        <property name="java-output-path" value="/dashboard/target/classes"/>
     </wb-module>
 </project-modules>
