+++ pom.xml
@@ -2,11 +2,11 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.lamminhthien</groupId>
-    <artifactId>struts2_crud_mybatis</artifactId>
+    <groupId>com.thuctap</groupId>
+    <artifactId>dashboard</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>war</packaging>
-    <name>struts2_crud_mybatis</name>
+    <name>dashboard</name>
 
     <properties>
         <struts2.version>2.5.22</struts2.version>
