+++ src/main/resources/log4j2.xml
@@ -8,7 +8,7 @@
     <Loggers>
         <Logger name="com.opensymphony.xwork2" level="info"/>
         <Logger name="org.apache.struts2" level="info"/>
-        <Logger name="com.lamminhthien" level="debug"/>
+        <Logger name="com.thuctap" level="debug"/>
         <Root level="warn">
             <AppenderRef ref="STDOUT"/>
         </Root>
