+++ src/main/resources/struts.xml
@@ -6,13 +6,13 @@
     <constant name="struts.devMode" value="true" />
     <package name="default" namespace="/" extends="struts-default">
         <action name="index"
-                class="com.lamminhthien.struts2_crud_mybatis.Index"
+                class="com.thuctap.dashboard.Index"
                 method="execute">
             <result name="success">/index.jsp</result>
         </action>
         
         <action name="create"
-                class="com.lamminhthien.struts2_crud_mybatis.Create"
+                class="com.thuctap.dashboard.Create"
                 method="execute">
             <result name="success">/index.html</result>
         </action>
